test: add regression test for https://github.com/django-cms/django-cms/issues/8108
